Day 6

+++ AdventOfCode2023/Days/Day.cs
@@ -10,11 +10,11 @@ public abstract class Day
         Console.WriteLine($"Part 2: {GetPart2Answer(false)}");
     }
 
-    public int GetPart1Answer(bool isTest) => isTest ? SolvePart1(GetPart1Lines()) : SolvePart1(GetLines());
-    public int GetPart2Answer(bool isTest) => isTest ? SolvePart2(GetPart2Lines()) : SolvePart2(GetLines());
+    public long GetPart1Answer(bool isTest) => isTest ? SolvePart1(GetPart1Lines()) : SolvePart1(GetLines());
+    public long GetPart2Answer(bool isTest) => isTest ? SolvePart2(GetPart2Lines()) : SolvePart2(GetLines());
 
-    protected abstract int SolvePart1(List<string> lines);
-    protected abstract int SolvePart2(List<string> lines);
+    protected abstract long SolvePart1(List<string> lines);
+    protected abstract long SolvePart2(List<string> lines);
     protected abstract int GetDayNum();
     private const string Path = @"F:\Documents\Code\AdventOfCode2023\AdventOfCode2023\Data\day";
     private List<string> GetLines() => File.ReadLines($@"{Path}{GetDayNum()}\data.txt").ToList();
