+++ AdventOfCode2023/Days/Day5.cs
@@ -6,7 +6,7 @@ public class Day5 : Day
 {
     protected override int GetDayNum() => 5;
 
-    protected override int SolvePart1(List<string> lines)
+    protected override long SolvePart1(List<string> lines)
     {
         var maps = GetMaps(lines);
 
@@ -27,7 +27,7 @@ public class Day5 : Day
         return (int)seedLocation.Item2;
     }
 
-    protected override int SolvePart2(List<string> lines)
+    protected override long SolvePart2(List<string> lines)
     {
         var reversedMaps = GetMaps(lines).Reverse().ToArray();
